Create a new SaveOrUpdate method on NomisUsers
We want to address a race condition that appeared today where a user was able to execute this method twice in quick succession[1].

When used this new method will allow the NomisUserService to recover gracefully from the violation of the uniqueness validation on username, meaning the user will not receive an error.

The key part of this method is the `ON CONFLICT (nomis_username) DO UPDATE SET` line and beyond. We’re attempting to make this request idempotent by allowing only this method to bypass the database validation and perform an update instead.

[1] https://mojdt.slack.com/archives/C05FN6Y2CSV/p1715087044061189