Move SQL logic from application to new database VIEW.
This commit creates a new Cas2ApplicationSummaryEntity dataclass and two new database VIEWS.  The first VIEW selects all application summary data from the relatve database columns, the second VIEW then selects unexpired views based on conditional release dates (as per existing logic).  The purpose of this is to remove the complex SQL logic from the application layer to the database layer.  The new dataclass is bound to columns in the second VIEW using JPA and therefore allows query methods to be used to extract and filter the necessary records.

Since the Cas2ApplicationSummary interface is no longer required this has been removed, along with a legacy function getStatusFromSummary() in ApplicationsTransformer which is no longer used.