Fix status for approved RequestForPlacement
Previously the status of ‘awaitingMatch’ was returned for a submitted but not assessed placement_application. This commit instead returns ‘requestSubmitted’.

This commit also introduces a ‘request unsubmitted’ status. Whilst this is logically never returned in any API call (because we ‘hide’ unsubmitted placement_applications), it allows us to ensure that our deriveStatus() is complete wrt. the state transition diagram