ensure assessments are created for status updates
now that we are adding an assessment to all status updates, we need to ensure
any seeded data adds assessments to status updates, otherwise
we would have data on dev/test being seeded without them

This ensure an assessment is created before the status updates,
and then added to the Application entity, because it was not
updating when using the Application to create new status updates.

I've included an integration test check to ensure the auto script works,
but doesn't look like there are currently tests for the seed job... I think a bit out of scope
for this change to add them, and as they are only for dev/test I don't think it's
an emergency to add them right now.