Add a Cas1EmailNotificationService
This is in preparation for setting application-specific reply-to email addresses for CAS1 emails.

All Cas1 Email Services will be updated to use this service.

A non-nullable ‘application’ argument will then be added to email the functions, and this will be used to determine the CRU email address (via application.cru.emailAddress) and set as the reply-to value.