Change state when a decision is made for an application with an open clarification note
This fixes the bug introduced in #1805 which prevents any state
transition when the current state is `REQUESTED_FURTHER_INFORMATION`.
This prevents an update to the assessment from changing the status, but
also prevents a decision being made changing the status, which is
incorrect.

A test has been introduced that verifies the expected behaviour as
outlined in #1805 to make sure that this change does not cause the
original bug to reappear.