Add framework to stream report data to output
Currently, when constructing a report we load the entire report content into memory at multiple points:

1. In a List<A>, where A is the repository row definition
2. Then mapped into a List<B>, where B is the service layer row definition
3. Then written to an XLSX document

This commit adds a framework to ‘stream’ report output directly to its intended target. In this first version there is only the option to write to XSLX, which still results in the report being loaded into memory once (see step 3 above)

If CSV was used instead of XSLX, we could directly stream the CSV document to the servlet outputstream, ensuring we never load the full report into memory.

This change has been made in a way that it can easily be applied to all existing reports. SQL would need to be modified as follows:

1. Modify the output column names to use lower_cased_underscore_naming instead of camel case. This is required because postgres will always lowercase column names in the output. The new reporting framework will convert these back to camel case when outputing to XLSX
2. Because all values are written as String or Int types into XSLX, if specific date formatting is required this should be achieved by formatting dates to varchar/text types in the SQL itself