Request Logger was breaking binary downloads
This commit fixes an issue in the HttpLoggingFilter that was breaking download of binary files that were streaming to the Servlet’s OutputStream. This is only an issue in environments where request/response logging is enabled (i.e. local deploys)

Because binary files are handled asynchronously outside of the servlet filter thread execution, we have to add some additional logic to listen for this asynchronous processing to complete before writting the contents of the ‘intercepted’ outputstream back to the caller