Demonstrate LEGACY_MANAGER can access new /cas1/ out-of-service-beds endpoints
The existing lost bed functionality has been moved to new endpoints:

/cas1/premises/{premisesId}/out-of-service-beds/*

which are handled by the new CAS1 `LostBedsController`.

In this commit we demonstrate that the new LEGACY_MANAGER role can use these
endpoints.

Similar to the `PremisesController`, we have a specific check on the
permitted roles within the `LostBedsController`. I think this is redundant as
the same controls are enforced by `UserAccessService.currentUserCanManagePremisesLostBeds()`.