add Service function to abandon Application
We want to allow the user to mark an Application as abandoned.
When an application is abandoned, it should have its answer data
wiped (the `data` field on an Application Entity).
This adds an ApplicationService method for the logic needed to
abandon an Application.
We check whether it is possible to mark the application as abandoned
- whether its already been submitted/abandoned/ whether
it belongs to another user/whether it exists - and then adds the
`abandonedAt` timestamp and sets `data` to `null`.