Add endpoint to abandon CAS2 Applications
We want to allow users to flag in progress
applications as abandoned.

This adds the functionality for Referrers to make
a PUT request to
`applications/{applicationId}/abandon`. We
decided to add a custom endpoint for this and
break from the traditional REST pattern, because
of the action that is required - it's not a hard
DELETE request because the Entity will remain in
our db, but its also not really an UPDATE request
because the entity is being deleted from the UI.
Having a custom end point also hopefully makes the
endpoint's purpose clear to the consumer, with a
single responsibility.