Add `abandonedAt` value to CAS2 Applications
We want to allow users to flag in progress
applications as abandoned. In order to do this we
will add an `abandonedAt` value to CAS2
Applications - similar to the `submitted_at`
value, this timestamp will determine whether an
application has been abandoned or not.

This adds the migration to add the new column and
the field on the Application Entity.