Support content-disposition when streaming files
Add a new generateStreamingResponse function that allows the caller to define the filename which is then written to the content-disposition header. The function also allows the content type to be defined, in preparation for adding CSV support.

This function has been added as we’re going to move setting report filenames out of the CAS1 UI and into the API