Use jsonb for domain_events.data
The domain_events.data column stores JSON data. We use data stored in this column in a lot of reports, and it has been noted that these reads are causing notable performance hits (e.g. it can add seconds to query execution times).

This commit updates domain_events.data to use the ‘jsonb’ type instead of the ‘json’ type. This should significantly improve read speed, in exchange for a slight performance penalty on inserts. As we typically only insert a handful of domain events per thread of execution, if any at all, this is an acceptable trade off.