Don’t use Application JPA Callback for AppStatus
Before this commit we used JPA Listeners to manage updating the application status when a Application Entity is persisted or updated.

Not only does this make comprehension of the code difficult, it technically isn’t allowed according to the JPA Spec, and has caused us issues when attempting to update to Spring 3 / Hibernate 6:

“In general, the lifecycle method of a portable application should not invoke EntityManager or Query operations, access other entity instances, or modify relationships within the same persistence context. A lifecycle callback method may modify the non-relationship state of the entity on which it is invoked.”

This commit calls the ApplicationListener class directly as an intermediate step towards using a state machine to manage application state

As part of this change the order that the ‘listeners’ were invoked in has changed, leading to the AssessmentListener overwriting the Withdrawn Application Status. AssessmentListener has been updated to correct this, and the WithdrawalTest updated to assert on application state after withdrawal