Add Clock to Spring Context
This allows us to control the system clock during integration tests.

A MutableClock implementation is provided in the integration test Application Context to support setting the current time.

This commit also uses the clock to set the application submission time, as we need to control this for Application Report Tests