Add CAS1 CRU_MEMBER role
This role will be used for members of the national CRU team. We'll
use it to grant access to

- managing out-of-service bed across all areas / premises
- withdrawing placements
- creating a manual adhoc placement

These are feature which live within the "manage" section of the service
but aren't for the use of AP managers or AP staff.

Currently the WORKFLOW_MANAGER role has been used for this purpose
but this is confusing an incorrect as WORKFLOW_MANAGER is a role
for CRU team members who can (re)allocate tasks: assigning assessors.

Note that `UsersController.transformApiRole()` and `UserServiceUtils.transformUserRoles()`
are both not violating our `detekt` complexity rules so we may want to
break down these functions.