Introduce JPA entity for out-of-service bed revisions
To provide rich information about the scope of each update, when a new
`Cas1OutOfServiceBedRevisionEntity` is created a comparison is done
against the most recent existing one for the start date, end date,
reason ID, reference number, and notes, which is then packed into the
`change_type` column as bit flags. This is unpacked into a set of
`Cas1OutOfServiceBedRevisionType` values when marshalling the API
response. This approach avoids having to perform the comparison for each
revision whenever an out-of-service bed record is requested.

Other options to persist this were considered:
- A tag table could have been defined with `id` and `value` columns and
  a `UNIQUE` constraint across both. This would allow a list of change
  types for each revision without duplicates, but has the downside of
  introducing an additional join.
- A `TEXT` column could have been used with comma-separated values,
  which could then be parsed following a `String.split()` call. However,
  this is essentially a less space-efficient version of the bit flags
  approach, and introduces the issue of how to handle unrecognised data.
- Several `BOOLEAN` columns could have been used to hold the flags.
  This is easier to use in `WHERE` clauses in queries, but filtering the
  history by revision type seems like an unlikely use case and also
  requires a migration to update all existing rows whenever a new change
  type is added.

Using a `BIGINT` as a bit flags means that up to 64 change types can
introduced in a backwards compatible way, as it's the responsibility of
the application to pack and unpack this value.

Currently, the flags are laid out like so:
```
| 63 ... 08 | 07 | 06 | 05 | 04    | 03     | 02       | 01       | 00         |
| <unused>                 | notes | reason | ref. no. | end date | start date |
```

A value of `0L` represents no change, and should only be used for
initial revisions.