Add Updated CAS1 Application Raw Data Report
This commit provides a rewrite of the CAS1 RAW Applications Data Report, adding and removing several fields and outputting the report as a CSV.

The commit also adds the Cas1SimpleApiClient in an attempt to make tests interacting with the API to setup test data more readible.

Instead of replacing the original report, this report is added as a new version of the original report. This allows us to feature flag the availability of the report on the UI as we’re not currently in a position to backfill required data in production.