CAS1: Revert change to QCode in 'Test AP 10'
In [8ef42b53cb3e4a6a3c6120a5233d5e09ec0ddbea][] we tweaked the
seeding of CAS1 Premises in the local and dev environments to
get the E2E tests passing locally. Unfortunately this has broken
the implementation in the dev environment which relies on a Premises
with the name:

- "Test AP 10" and
- the QCode Q710

I'm now attempting to add a premises with this Q710 qcode (Delius Hostel ID)
to the AP-and-Delius seeding in their local environment (confusingly labelled
'dev').

So this commit will:

- allow the CAS1 `e2e/tests/manage.spec.ts` to pass NOW in our deployed
  "dev" env

- hopefully allow that test to pass locally SOON, once the necessary seeding
  has been added to AP-and-Delius

[8ef42b53cb3e4a6a3c6120a5233d5e09ec0ddbea]:
https://github.com/ministryofjustice/hmpps-approved-premises-api/commit/8ef42b53cb3e4a6a3c6120a5233d5e09ec0ddbea