Backfill CAS1 Application Timeliness
This commit provides a SQL migration to populate the approved_premises_application.timeliness value for any application where it isn’t currently set. This is ensure the value is always available in reports.

This value was only populated in the database as-of 2024-03-20. This migration will only update applications where the value is currently null (i.e. from before 2024-03-20).

The logic used to determine the timeliness values matches that used in the UI (noticeTypeFromApplication.ts), which is applied as part of the application submission logic.