Add Placement Matching Outcomes Report v2
This commit provides a revised version of the placement matching outcomes report.

This is built on top of the Request for Placement V2 Report, using that report SQL as a CTE to ‘drive’ the report data. To support this some additional ‘internal’ columns have been added to the Request for Placement V2 report that can be used to join data. These will be filtered out when returning the reports to users.