Move remaining CAS1 user seeding to AutoScript()
JIMSNOWLDAP is still created in `R__1_create_dev_users.sql` but we add
the CAS1 roles to that user here.

That user is also needed by CAS3 who continue to assign some CAS3 roles
in that historic SQL migration.

We add a `documentation` property to the `SeedUser` data class to help
us explain the purpose of each user seeded.