Fix traces sampling rate for Sentry transaction profiling
It was brought to our attention that the CAS API service
has been using an excessive amount of the cloud platform's
overall quota.

We reduce the *traces* sampling rate `SENTRY_TRACES-SAMPLE-RATE`
from 1.0 to 0.01 as per the [Cloud Platform guidance][].

We leave the `SENTRY_SAMPLE-RATE` at 1.0 as we wish to [record
every error at Sentry][], for now.

See [Sentry docs on SpringBoot][] for further information.

[Sentry docs on SpringBoot]:
https://docs.sentry.io/platforms/java/guides/spring-boot/configuration/sampling/#setting-a-sampling-function

[Cloud Platform guidance]:
https://user-guide.operations-engineering.service.justice.gov.uk/documentation/services/sentry.html#transaction-sampling

[record every error at Sentry];
https://docs.sentry.io/platforms/java/guides/spring-boot/configuration/sampling/#sampling-error-events