Add ‘Update NOMS Number’ Seed Job
This commit adds a seed job that allows us to update the NOMS number associated with a given CRN on applicationa and bookings. This is required when the NOMS number is updated in delius, which is rare, but can happen.

This helps reduce the number of 404 errors we see when the inmate cache is refreshed because we’re removing usage of invalid NOMS numbers.

f