Add config to exclude NOMS from cache refresh
When a NOMS number can’t be found in the prisons-api, we receive a 404. If this occurs multiple times when refreshing inmate details in the cache (via InmateDetailsCacheRefreshWorker), we start to receive Sentry errors that this has occurred.

It’s possible that the resolution to the issue (typically Delius or NOMIS having the wrong NOMS number for the CRN) may take days or weeks. In the meantime we no longer need to see the cache refresh warnings, and we should stop attempting to refresh the inmate details as it will fail.

This commit provides a mechanism to exclude NOMS from inmate details cache refresh. Entries can be added to the cache_refresh_exclusions_inmate_details table using SQL