Enforce presence of OutOfServiceBedRevision.notes
We will be migrating existing LostBed records to the new OutOfServiceBed
models. As `LostBed.notes` are nullable we won't enforce the presence of
notes on the `OutOfServiceBedRevision` entity yet. Instead we use
a validation step within the `OutOfServiceBedService`.

Once the new OOSB system is released and the `LostBed` data has been
migrated we can make `OutOfServiceBedRevision.notes` non-nullable