Add PR ID to CAS1 Booking Made metadata
This commit adds the placement_request ID to the APPROVED_PREMISES_BOOKING_MADE domain event metadata, and backfills it for existing events.

This will only backfill the most recent booking made for a placement request because the value of placement_request.booking_id overwritten when rebooked (e.g. after a booking cancellation), but for reporting purposes this is adequate