Revert removal of generated open api YML files
The YML files in the folder are generated by the openApiGenerate gradle task.

Whilst it's generally bad practice to check in code that can be generated on the fly, we currently required these files to available in the git hub repository to allow the user interface projects to bull the latest 'compiled' open api specs for the project.

A previous commit removed these generated files as we hadn’t appreciated how they were being used.