Use incrementing num for integration test db name
Before this commit a random name was given to the database generated for an integration test. This made it difficult to find the database created for the most recent execution of the integration tests.

This commit shortens the fixed part of the database name and suffixes it with an epoch milli, a number that will always increment. This ensures that the database used for the latest execution of tests will always appear last in the list of the available databases