Apply LAO Qualification to CAS1 ‘All Applications’
Before this change users with the LAO Qualification could not navigate to an LAO application when using the ‘all applications’ screen on the Apply tab.

This commit allows users with the LAO Qualification to view LAO offenders names on the ‘all applications’ screen, and provides a link to click through to the application, as they can do elsewhere in the application.

The previous behaviour was (presumably) in place because the LAO qualification is not sufficient to apply for a new application, and this logic continued to the ‘All Applications’ screen which is also on the Apply Tab. Despite being on the Apply Tab, the all application screen is really a general purpose application search so loosening of restrictions here is fine.

This addresses an issue where users could not create appeals for LAO applications as there was no straight forward path to view the application.