Allow ASSESSORs to assess placement_applications
This commit allows users with the role ASSESSOR to be considered when allocating/reallocating assessments by including them in the task’s users list.

This commit also changes permissions on various endpoints required to complete the asessment of a placement application, ensuring users with the ASSESSOR role can access them.