Make probation_region.ap_area_id nullable
ap_areas should be related to a geographic region. We currently have a ‘National’ ap_area that doesn’t make logical sense because ‘national’ isn’t a geographic region. We are currently working on removing this ap_area.

We also have a ‘National’ probation_region linked to the ‘National’ ap_area. We want to retain this probation region. To support this, we need to make the probation_region.ap_area_id relationship nullable, subsequently setting it to null for the ‘National’ probation_region (and _only_ the ‘National’ probation region)

There are two contexts in which probation_region is used in the code to consider:

* When determing a user’s AP Area. In this case the code is already ignoring the probation_regions.ap_area_id relationship and instead determines the ap_area by looking at the user’s team code(s). See Cas1ApAreaMappingService for more information
* When determining the premises AP Area. There are no premises linked to the ‘National’ probation region, so this value being null for this specific region should not cause any issues

This commit makes the probation_region.ap_area_id column null in the DDL and Kotlin.