Simplify Cas1ApAreaMappingService
Prior to this commit the Cas1ApAreaMappingService used the Delius User’s Probation Area Code to determine if the team code should be used to determine the user’s ap_area, instead of the user’s probation_region.ap_area_id mapping. In this case applicable probation area codes were N43, N41, XXX. This was essentially reasserting the mapping defined in probation_area_probation_region_mappings, which identified these 3 codes as belonging to the ‘National’ probation_region.

Given that the prior commit has now ‘nulled’ the probation_region.ap_area_id value for the ‘National’ probation region, we can leverage this to simplify the Cas1ApAreaMappingService. The code now falls back to the team codes to determine a user’s ap_area if the probation_region.ap_area_id value is null. This makes logical sense as using team codes is a ‘fall back’ for when a probation region doesn’t directly map to an ap area, which is currently only the case for the national probation region.

This change helps remove duplication of mappings, and simplify testing.