Add indexes to optimise task workload query
This commits adds indexes to ‘task’ types in database to optimise any query that determines which tasks are allocated to a given user. We include ‘reallocated_at’ in this index because this is always checked alongside the allocated_to_user_id (because if reallocated, these tasks are effectively inactive and should be ignored).