Add `JpaRepository` extension to get entities by IDs in given order
The `CrudRepository.findAllById` method does not guarantee the order of
the elements returned. This commit adds the `findAllByIdOrdered` method
as an extension to the `JpaRepository` interface, which guarantees that
the returned entities will be ordered according to the provided list of
IDs, i.e. that the first entity will have as its ID the first element of
the input list, the second entity will have the second element of the
input list as its ID, and so on.

This is achieved by constructing a map from the results returned by the
original `findAllById` method, using the ID as the key, and then mapping
the input list of IDs to the result using the map.

To do this in a generic way requires using reflection. This can be
written manually, however it is possible to get the ID using a
`JpaEntityInformation` instance. To do this requires an active
`EntityManager`, which is difficult to get an instance of outside of a
Spring context. Therefore, the `RepositoryUtilHelper` class acts as a
bridge to the static context - as a Spring component, the
`EntityManagerFactory` is autowired on construction, and the
`@PostConstruct` method registers itself to the static `lateinit`
instance - allowing it to be accessed by the utility functions.