Fix incorrect sorting of out-of-service beds in real-world usage
It has been discovered through the end-to-end tests that sorting is not
applied to out-of-service beds when the sorting controls on the
dashboard are used. Unfortunately, the integration test for the `GET
/cas1/out-of-service-beds` endpoint cannot replicate this issue, which
makes it difficult to identify the cause.

However, the underlying issue is that the `findAllById` method
explicitly does not guarantee the order of the returned results. The
order is therefore dependent entirely on how the results are returned
from the database, which may be non-deterministic.

To resolve this, a new utility method has been introduced to wrap the
`findAllById` method while guaranteeing the order of the returned
elements, which solves the issue.