Use latest open api generator plugin
This is in preparation for the spring boot 3 upgrade

There appears to be a bug in the generator that is setting a value of type ‘Any’ on a subclass to be optional e.g. Any?, despite it being marked as required on the superclass. Attempts to fix this by setting required on the subclass or changing how inheritence works in the open api has not helped. As a work-around, we’ve marked the field as optional in the open api definition. I supect upgrading the open api schema version in our yml files may help resolve this, but this would most likely impact on the UI clients too.