Deprecate OffenderDetailSummary
OffenderDetailSummary is from the community api model. Instead we should be using [CaseSummary] which is the model provided by ap-and-delius, which has been tailored for CAS usage. Currently we use both which is confusing and duplication of effort.

See APS-1085 for some notes on how to replace usage of this class in some circumstances. Note that OffenderDetailsUtils provides functions to convert between these two types