Add Cas1SpaceSearchRequirements API model allow searching for multiple ApType and Gender
When we book a space in an AP it is for a specific type and
particular gender, as per `Cas1SpaceBookingRequirements` --
but when searching for available spaces we allow the
matcher to search for availability in multiple types of premises
and in premises for both genders.

To do this we introduce a `Cas1SpaceSearchRequirements` API model.