Search for spaces using 'space characteristics'
Now that we have modelled "space characteristics" (a combination
of premises characteristics and room characteristics) and are
using these to make a space booking with lists of:

- `essentialCharacteristics`, and
- `desirableCharacteristics`

It's clear that the concepts of `needCharacteristics` and
`riskCharacteristics` should be UI concerns only. We are
already intimately involved with these characteristics on
the UI, e.g. in https://github.com/ministryofjustice/approved-premises-ui/blob/7d7041a6a99934b89afc6834928ad4f7eb0fa360/server/utils/placementCriteriaUtils.ts