Add `zipBy` utility method
This method joins two `Iterable`s according to some ID function such
that each element of the result `Iterable` is a `Pair` where both
elements have the same ID. This function maintains the order of the
iterable on which this method is called.

This is useful for a situation where the first query returns a list of
IDs in some order, and a second query takes in that list of IDs as a
parameter and provides some augmented data, but does not guarantee any
order, and the desired result returned to the caller associates results
of both queries together.