updated code for readability and to include extra check on user being found
updated test

updated tests to return mock call from delius

updated tests to match transformer

updated transformer to match the open api spec - deserialization expects the service name to be CAS3, not temporary-accommodation

added new integration test

added check to make sure user is not null.

refactored to update user details if the user has not been created as part of the get request.

added createdOnGet flag, so we can tell if the user has been created as part of the retrieval.

overloaded method to allow entity or id to passed in, as we already have the entity in the controller.

changed return type of updateUserFromCommunityApiById