Use sealed hierachy for GetUserResponse
This forces the reciver to consider all possible outcomes and should make the possible reponse types clearer, allowing us to use ‘when’ to check all outcomes. It also removes the need to use user!! on success responses.

This commit also updates the response from updateUserFromCommunityApi, removing the AuthorisableActionResult wrapper as previously we always returned the Success outcome, so it was just complicating the response

This commit also returns a 404 on calls to users/{id} if a staff record isn’t found. Previously it would have returned a 500 (via null pointer exception)

This still leaves a slightly awkward response from ‘updateUserFromCommunityApiById’ of AuthorisableActionResult<GetUserResponse>, to cover the ‘user entity not found’ case. I propose we refactor this code such that the calling code retrieves the user entity and then passes it into updateUserFromCommunityApi. The calling code can then deal with the user entity not found scenario itself.