Remove throwExceptionOnStaffRecordNotFound=true
This commit removes any call to getExistingUserOrCreate where throwExceptionOnStaffRecordNotFound=true. This is possible because GetUserResponse is now being handled by all callers to this function via a ‘when’ statement, meaning they are all handling the case when a staff record isn’t found themselves.

The overloaded version of getExistingUserOrCreate that throws an exception when staff record isn’t found has subsequently been deprecated, as all callers should be dealing with GetUserResponse directly to determine what to do if a staff record isn’t found